+++ keychain_default.go
@@ -13,16 +13,16 @@ func keychainFetcher(serviceLabel string) string {
 	}
 	return secret
 }
-func keychainWriter(serviceLabel string, secret string) {
+func keychainWriter(serviceLabel, secret string) {
 	err := keyring.Set(serviceLabel, serviceLabel, secret)
 	if err != nil {
 		panic(err)
 	}
 }
 
-func keychainDeleter(serviceLabel string) {
-	err := keyring.Delete(serviceLabel, serviceLabel)
-	if err != nil {
-		panic(err)
-	}
-}
+// func keychainDeleter(serviceLabel string) {
+// 	err := keyring.Delete(serviceLabel, serviceLabel)
+// 	if err != nil {
+// 		panic(err)
+// 	}
+// }
